% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_aux.R
\name{reg_aux}
\alias{reg_aux}
\title{Auxiliary regression of all covariates on one (or more) variable of interest}
\usage{
reg_aux(...)
}
\arguments{
\item{object}{output from regression}

\item{var_main}{a character, name of the variable of interest}

\item{var_controls}{a character, name of the covariates}
}
\value{
An object of the same type
}
\description{
Given a regression of y on x and covariates, estimate the correlation between all covariates and the variable x
}
\details{
This is a generic function: methods can be defined for it directly
or via the \code{\link{Summary}} group generic. For this to work properly,
the arguments \code{...} should be unnamed, and dispatch is on the
first argument.
}
\examples{
  model_lm <- lm(y ~ lag.quarterly.revenue + price.index + income.level + market.potential, data=freeny)
  res_lm <- reg_aux(object=model_lm, var_main = "lag.quarterly.revenue")
}
